More fine grained selection of cluster roles
Add new aliases to extract either the resources, verbs, non resource
urls, or api groups out of a role. This depends on a new indexing
function index_select where the index is a number, and the suffix of the
field name indidcates what property to extract out of the ith object in
the array.