Change exec/attach rules to work for all pods
Although the description of the rule Attach/Exec Privileged Pod claimed
to work for privileged pods, it actually worked for all pods, as knowing
if a named pod was privileged will require some state management within
falco, which isn't supported yet.

Change the rules to note this with comments talking about what is
required to support privileged attach/execs.